+++ arcane/src/arcane/accelerator/CMakeLists.txt
@@ -15,9 +15,9 @@ arcane_accelerator_add_source_files(MemoryCopier.cc CommonUtils.cc)
 target_link_libraries(arcane_accelerator PUBLIC
   arcane_core
   arcane_accelerator_core
-  "$<TARGET_NAME_IF_EXISTS:arcane_accelerator_cuda>"
-  "$<TARGET_NAME_IF_EXISTS:arcane_accelerator_hip>"
-  "$<TARGET_NAME_IF_EXISTS:arcane_accelerator_sycl>"
+  "$<TARGET_NAME_IF_EXISTS:Arccore::arccore_accelerator_cuda>"
+  "$<TARGET_NAME_IF_EXISTS:Arccore::arccore_accelerator_hip>"
+  "$<TARGET_NAME_IF_EXISTS:Arccore::arccore_accelerator_sycl>"
 )
 
 # Cette cible est optionnelle pour éviter que le code n'utilisant
