+++ arcane/src/arcane/accelerator/cuda/runtime/CMakeLists.txt
@@ -15,10 +15,12 @@ arcane_add_library(arcane_accelerator_cuda_runtime
   FILES ${ARCANE_SOURCES}
   )
 
-target_link_libraries(arcane_accelerator_cuda_runtime PRIVATE arcane_accelerator_cuda)
+target_include_directories(arcane_accelerator_cuda_runtime PRIVATE $<BUILD_INTERFACE:${Arcane_SOURCE_DIR}/src>)
+
+target_link_libraries(arcane_accelerator_cuda_runtime PRIVATE Arccore::arccore_accelerator_cuda)
 
 target_link_libraries(arcane_accelerator_cuda_runtime PUBLIC
-  arcane_cuda_compile_flags
+  arccore_cuda_compile_flags
   CUDA::cudart
   CUDA::cuda_driver
 )
@@ -46,7 +48,7 @@ endif()
 add_library(arcane_accelerator_cuda_testlib Test.cu Test2.cc TestCpp.cc)
 arcane_accelerator_add_source_files(Test2.cc)
 set_target_properties(arcane_accelerator_cuda_testlib PROPERTIES CUDA_SEPARABLE_COMPILATION ON)
-target_link_libraries(arcane_accelerator_cuda_testlib PUBLIC arcane_accelerator arcane_core arcane_cuda_compile_flags)
+target_link_libraries(arcane_accelerator_cuda_testlib PUBLIC arcane_accelerator arcane_core arccore_cuda_compile_flags)
 
 add_executable(arcane_accelerator_cuda_test TestMain.cc)
 target_link_libraries(arcane_accelerator_cuda_test PUBLIC arcane_accelerator_cuda_testlib arcane_accelerator_cuda_runtime)
