+++ arcane/src/arcane/accelerator/cuda/runtime/CudaAcceleratorRuntime.cc
@@ -11,7 +11,7 @@
 /*---------------------------------------------------------------------------*/
 /*---------------------------------------------------------------------------*/
 
-#include "arcane/accelerator/cuda/CudaAccelerator.h"
+#include "arccore/accelerator_native/CudaAccelerator.h"
 
 #include "arccore/base/CheckedConvert.h"
 #include "arccore/base/FatalErrorException.h"
