+++ arccore/src/accelerator_native/arccore/accelerator_native/CMakeLists.txt
@@ -1,0 +1,8 @@
+﻿include (${CMAKE_CURRENT_SOURCE_DIR}/native_${ARCCORE_ACCELERATOR_RUNTIME}.cmake)
+
+# ----------------------------------------------------------------------------
+# Local Variables:
+# tab-width: 2
+# indent-tabs-mode: nil
+# coding: utf-8-with-signature
+# End:
