+++ arccore/src/accelerator_native/arccore/accelerator_native/native_cuda.cmake
@@ -1,28 +1,27 @@
 ﻿# ----------------------------------------------------------------------------
 # ----------------------------------------------------------------------------
 # Backend Arcane pour CUDA
-# Cette composante est interne à Arccore/Arcane
 
-set(ARCANE_SOURCES
+set(ARCCORE_SOURCES
   CudaAccelerator.cc
   CudaAccelerator.h
   )
 
-arcane_find_package(CUDAToolkit REQUIRED)
+find_package(CUDAToolkit REQUIRED)
 
-if ((ARCANE_HAS_CXX23) AND (CMAKE_CUDA_COMPILER_VERSION VERSION_LESS 13.0))
+if (ARCCORE_HAS_CXX23 AND (CMAKE_CUDA_COMPILER_VERSION VERSION_LESS 13.0))
   message(FATAL_ERROR "CUDA <=12 doesn't support C++23. Add -DARCCORE_CXX_STANDARD=20 to the configuration")
 endif()
 
 # Créé une cible interface pour propager les options de compilation
 # communes pour la compilation CUDA
 
-add_library(arcane_cuda_compile_flags INTERFACE)
-add_library(arcane_cuda_build_compile_flags INTERFACE)
+add_library(arccore_cuda_compile_flags INTERFACE)
+add_library(arccore_cuda_build_compile_flags INTERFACE)
 
-option(ARCANE_CUDA_DEVICE_DEBUG "If True, add '--device-debug' to cuda compiler flags" OFF)
+option(ARCCORE_CUDA_DEVICE_DEBUG "If True, add '--device-debug' to cuda compiler flags" OFF)
 set(_CUDA_DEBUG_FLAGS "-lineinfo")
-if (ARCANE_CUDA_DEVICE_DEBUG)
+if (ARCCORE_CUDA_DEVICE_DEBUG)
   set(_CUDA_DEBUG_FLAGS "--device-debug")
 endif()
 
@@ -31,13 +30,13 @@ if (CMAKE_CUDA_COMPILER_ID STREQUAL "Clang")
   # de debug car cela provoque une erreur de compilation
   # (voir https://github.com/llvm/llvm-project/issues/58491)
   # Cela sera peut-être corrigé avec la version 19 de clang
-  target_compile_options(arcane_cuda_compile_flags INTERFACE
+  target_compile_options(arccore_cuda_compile_flags INTERFACE
     "$<$<COMPILE_LANGUAGE:CUDA>:-Xarch_device>"
     "$<$<COMPILE_LANGUAGE:CUDA>:-g0>"
   )
 else()
   # Compilateur CUDA classique (NVCC ou NVHPC)
-  target_compile_options(arcane_cuda_compile_flags INTERFACE
+  target_compile_options(arccore_cuda_compile_flags INTERFACE
     "$<$<COMPILE_LANGUAGE:CUDA>:--extended-lambda>"
     "$<$<COMPILE_LANGUAGE:CUDA>:--expt-relaxed-constexpr>"
     "$<$<COMPILE_LANGUAGE:CUDA>:-g>"
@@ -45,27 +44,23 @@ else()
     "$<$<COMPILE_LANGUAGE:CUDA>:cross-execution-space-call>"
   )
 
-  target_compile_options(arcane_cuda_build_compile_flags INTERFACE
+  target_compile_options(arccore_cuda_build_compile_flags INTERFACE
     "$<$<COMPILE_LANGUAGE:CUDA>:${_CUDA_DEBUG_FLAGS}>"
   )
 endif()
 
-install(TARGETS arcane_cuda_compile_flags EXPORT ArcaneTargets)
-install(TARGETS arcane_cuda_build_compile_flags EXPORT ArcaneTargets)
+install(TARGETS arccore_cuda_compile_flags EXPORT ArccoreTargets)
+install(TARGETS arccore_cuda_build_compile_flags EXPORT ArccoreTargets)
 
-arcane_add_library(arcane_accelerator_cuda
-  INPUT_PATH ${Arcane_SOURCE_DIR}/src
-  RELATIVE_PATH arcane/accelerator/cuda
-  FILES ${ARCANE_SOURCES}
-  )
+arccore_add_component_library(accelerator_native
+  LIB_NAME arccore_accelerator_cuda
+  FILES ${ARCCORE_SOURCES}
+)
 
-# Pour que le runtime associé trouve les fichiers d'en-tête de cette composante
-target_include_directories(arcane_accelerator_cuda PUBLIC $<BUILD_INTERFACE:${Arcane_SOURCE_DIR}/src>)
-
-target_link_libraries(arcane_accelerator_cuda PUBLIC
-  Arccore::arccore_common
-  arcane_cuda_compile_flags
-  $<BUILD_INTERFACE:arcane_cuda_build_compile_flags>
+target_link_libraries(arccore_accelerator_cuda PUBLIC
+  arccore_common
+  arccore_cuda_compile_flags
+  $<BUILD_INTERFACE:arccore_cuda_build_compile_flags>
   CUDA::cudart
 )
 
@@ -77,18 +72,13 @@ target_link_libraries(arcane_accelerator_cuda PUBLIC
 # qui en dépende mais peut-être s'agit-il d'une limitation de 'nvcc')
 # Elle permet aussi de propager les options de compilation aux utilisateurs
 # de cette bibliothèque
-add_library(arcane_accelerator_cuda_impl STATIC
-  Reduce.cu
-#  CudaReduceImpl.h
-  )
-set_target_properties(arcane_accelerator_cuda_impl PROPERTIES CUDA_SEPARABLE_COMPILATION ON)
-set_target_properties(arcane_accelerator_cuda_impl PROPERTIES LINKER_LANGUAGE CUDA)
-target_link_libraries(arcane_accelerator_cuda_impl PUBLIC arcane_core arcane_cuda_compile_flags)
-target_link_libraries(arcane_accelerator_cuda_impl PUBLIC CUDA::cudart)
-
-# ----------------------------------------------------------------------------
-
-arcane_register_library(arcane_accelerator_cuda  OPTIONAL)
+# add_library(arccore_accelerator_cuda_impl STATIC
+#   Reduce.cu
+#   )
+# set_target_properties(arccore_accelerator_cuda_impl PROPERTIES CUDA_SEPARABLE_COMPILATION ON)
+# set_target_properties(arccore_accelerator_cuda_impl PROPERTIES LINKER_LANGUAGE CUDA)
+# target_link_libraries(arccore_accelerator_cuda_impl PUBLIC arccore_core arccore_cuda_compile_flags)
+# target_link_libraries(arccore_accelerator_cuda_impl PUBLIC CUDA::cudart)
 
 # ----------------------------------------------------------------------------
 # Local Variables:
